requirements.txt
# Core
typer==0.12.4
pydantic==2.9.1
pydantic-settings==2.4.0
rich==13.7.1
structlog==24.4.0

# Schema & YAML
jsonschema==4.23.0
ruamel.yaml==0.18.6

# Dev quality
ruff==0.6.9
mypy==1.11.1
black==24.8.0
isort==5.13.2
pytest==8.3.3
pytest-cov==5.0.0

# Patch tooling
unidiff==0.7.5

# Orchestrator deps
requests==2.32.3
PyYAML==6.0.2
